# internal/api/openapi.json
{
  "openapi": "3.2.1",
  "info": {
    "title": "WG-Guard API",
...
    "version": "1.0.0"
  },
  "servers": [{ "url": "/" }],
  "security": [{ "BearerAuth": [] }],
...
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "tags": ["users"], "summary": "Create a user", "operationId": "createUser",
        "description": "Idempotent with `Idempotency-Key`. When template_id names an enabled template, its technical terms override submitted limit, duration, start-policy and interface fields. Without a template, absent limits mean unlimited. This route creates no device; use /purchases for atomic customer-plus-first-device provisioning. Speed limits are independent per direction (kbps).",
        "security": [{ "BearerAuth": ["users.create"] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserCreate" }, "example": { "username": "alice", "duration_seconds": 2592000, "traffic_limit_bytes": 100000000000, "speed_limit_down_kbps": 100000, "speed_limit_up_kbps": null, "device_limit": 3 } } } },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "400": { "$ref": "#/components/responses/Invalid" }, "409": { "$ref": "#/components/responses/Conflict" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "get": {
        "tags": ["users"], "summary": "List users (cursor pagination)", "security": [{ "BearerAuth": ["users.read"] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "maximum": 500, "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["created_at", "username", "expires_at", "used"], "default": "created_at" } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "username", "in": "query", "schema": { "type": "string" }, "description": "Substring filter (wildcards are literal)" },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/UserStatus" } },
          { "name": "traffic_exceeded", "in": "query", "schema": { "type": "boolean" } },
          { "name": "enabled", "in": "query", "schema": { "type": "boolean" } },
          { "name": "expires_before", "in": "query", "schema": { "type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds." } },
          { "name": "expires_after", "in": "query", "schema": { "type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds." } },
          { "name": "created_before", "in": "query", "schema": { "type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds." } },
          { "name": "created_after", "in": "query", "schema": { "type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds." } },
          { "name": "template_id", "in": "query", "schema": { "type": "string" } },
          { "name": "interface_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Page of users", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserPage" } } } } }
      }
    },
    "/api/v1/users/bulk": {
      "post": {
        "tags": ["bulk"],
        "summary": "Bulk-create users (prefix + zero-padded index)",
        "security": [{"BearerAuth": ["users.bulk"]}],
        "description": "Creates only user records, no devices or customer links. Enabled templates override manual entitlement terms. Unknown fields are rejected; metadata, enabled, auto_devices and exact expiry are not bulk-create properties.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["count"], "properties": {"count": {"type": "integer", "minimum": 1, "maximum": 500}, "prefix": {"type": "string", "example": "gs"}, "start_index": {"type": "integer", "default": 1}, "width": {"type": "integer", "default": 3}, "duration_seconds": {"type": "integer", "description": "Duration in seconds: 86400 = 1 day; 2592000 = 30 days. The panel month shortcut means 30 days, not a calendar month."}, "traffic_limit_bytes": {"type": ["integer", "null"], "description": "Quota in exact bytes (B). 1 GB = 1000000000 B; 100 GB = 100000000000 B. 100 GiB = 107374182400 B and displays as 107.4 GB. null = unlimited; 0 is a finite zero-byte allowance.", "minimum": 0}, "speed_limit_down_kbps": {"type": ["integer", "null"], "description": "Download (server to client) limit in decimal kilobits/s (1 Kbps = 1000 bits/s). 100000 = 100 Mbps = 12.5 MB/s in the panel speed form. null = unlimited.", "minimum": 1}, "speed_limit_up_kbps": {"type": ["integer", "null"], "description": "Upload (client to server) limit in decimal kilobits/s (1 Kbps = 1000 bits/s). 20000 = 20 Mbps = 2.5 MB/s in the panel speed form. null = unlimited.", "minimum": 1}, "device_limit": {"type": ["integer", "null"], "description": "Maximum registered device/peer configuration count, including disabled devices. null = unlimited; zero is invalid. This does not allocate devices, detect physical hardware or count simultaneous connections. Lowering it does not delete existing devices. Use purchase device_count to provision a specific number of ready configurations.", "minimum": 1}, "template_id": {"type": ["string", "null"], "description": "Optional enabled technical template ID. On create, saved quota, duration, start policy, device cap, speeds and interface override manual entitlement fields. Omit/null for manual terms."}, "interface_id": {"type": ["string", "null"], "description": "Connection profile ID, not a name such as awg0. Omit/null for automatic selection when creating a device. A user PATCH does not move existing devices."}, "start_policy": {"$ref": "#/components/schemas/StartPolicy"}, "display_name": {"type": "string", "description": "Shared presentation label."}, "note": {"type": "string", "description": "Shared internal text; JSON \\n preserves line breaks."}, "tags": {"type": "array", "items": {"type": "string"}, "description": "Shared tags as a JSON string array."}}, "additionalProperties": false}}}},
        "responses": {"201": {"description": "Created users + skipped count", "content": {"application/json": {"schema": {"type": "object", "properties": {"created": {"type": "array", "items": {"$ref": "#/components/schemas/User"}}, "skipped": {"type": "integer"}}}}}}}
      }
    },
    "/api/v1/users/bulk-action": {
      "post": {
        "tags": ["bulk"],
        "summary": "Apply one action to many users (per-item results)",
        "security": [{"BearerAuth": ["users.bulk"]}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["action", "user_ids"], "properties": {"action": {"type": "string", "enum": ["enable", "disable", "delete", "renew", "reset_traffic", "add_traffic", "update"]}, "user_ids": {"type": "array", "maxItems": 500, "items": {"type": "string"}}, "params": {"type": "object", "properties": {"reason": {"type": "string"}, "mode": {"type": "string", "enum": ["from_expiration", "from_now", "exact"], "description": "For action=renew; omitted defaults to from_expiration."}, "duration_seconds": {"type": "integer", "description": "Duration in seconds: 86400 = 1 day; 2592000 = 30 days. The panel month shortcut means 30 days, not a calendar month."}, "rx_bytes": {"type": "integer", "description": "Traffic received by the node from a client, in bytes (B); client upload. This is an amount, not a per-second rate."}, "tx_bytes": {"type": "integer", "description": "Traffic sent by the node to a client, in bytes (B); client download. This is an amount, not a per-second rate."}, "display_name": {"type": "string", "description": "Optional presentation label distinct from the username. PATCH empty string clears it; absent/null leaves it unchanged."}, "note": {"type": "string", "description": "Internal text; JSON \\n preserves line breaks. PATCH empty string clears it; absent/null leaves it unchanged. Never store credentials here."}, "traffic_limit_bytes": {"type": ["integer", "null"], "description": "Quota in exact bytes (B). 1 GB = 1000000000 B; 100 GB = 100000000000 B. 100 GiB = 107374182400 B and displays as 107.4 GB. null = unlimited; 0 is a finite zero-byte allowance.", "minimum": 0}, "speed_limit_down_kbps": {"type": ["integer", "null"], "description": "Download (server to client) limit in decimal kilobits/s (1 Kbps = 1000 bits/s). 100000 = 100 Mbps = 12.5 MB/s in the panel speed form. null = unlimited.", "minimum": 1}, "speed_limit_up_kbps": {"type": ["integer", "null"], "description": "Upload (client to server) limit in decimal kilobits/s (1 Kbps = 1000 bits/s). 20000 = 20 Mbps = 2.5 MB/s in the panel speed form. null = unlimited.", "minimum": 1}, "device_limit": {"type": ["integer", "null"], "description": "Maximum registered device/peer configuration count, including disabled devices. null = unlimited; zero is invalid. This does not allocate devices, detect physical hardware or count simultaneous connections. Lowering it does not delete existing devices. Use purchase device_count to provision a specific number of ready configurations.", "minimum": 1}, "exact": {"type": "string", "format": "date-time", "description": "For action=renew with mode=exact: RFC3339 expiry instant."}}, "additionalProperties": false}}, "additionalProperties": false}}}},
        "responses": {"200": {"description": "Per-item results", "content": {"application/json": {"schema": {"type": "object", "properties": {"results": {"type": "array", "items": {"type": "object", "properties": {"id": {"type": "string"}, "ok": {"type": "boolean"}, "error": {"type": "string"}}}}, "ok": {"type": "integer"}}}}}}},
        "description": "Per-user results; not an all-or-nothing transaction. params are action-specific: renew uses mode/duration_seconds/exact; add_traffic uses rx_bytes/tx_bytes to adjust charged usage, not quota; update accepts display_name, note, quota, independent speeds and device cap. Other single-user PATCH fields are not available here."
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "tags": ["users"], "summary": "Get a user", "security": [{ "BearerAuth": ["users.read"] }],
        "parameters": [{ "$ref": "#/components/parameters/UserID" }],
        "responses": { "200": { "description": "User", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "patch": {
        "tags": ["users"], "summary": "Partially update a user (username immutable; `null` clears a limit)", "description": "Changing template_id on an existing user changes its reference only; it does not reset the existing entitlement. Update quota/duration fields explicitly or queue a successor.", "security": [{ "BearerAuth": ["users.update"] }],
        "parameters": [{ "$ref": "#/components/parameters/UserID" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserPatch" } } } },
        "responses": { "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "delete": {
        "tags": ["users"], "summary": "Permanently delete an account and every device", "description": "One transaction removes the account, devices, customer link, chart history and queued next plan, releasing all its IP allocations. Username becomes available again. Former public peer keys remain as durable runtime removal intent until reconciliation succeeds. Audit and non-secret integration results retain their existing retention. Recovery requires a coordinated backup. Reseller tokens can delete only their own customers.", "security": [{ "BearerAuth": ["users.delete"] }],
        "parameters": [{ "$ref": "#/components/parameters/UserID" }],
        "responses": { "200": { "description": "Deleted" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
    "/api/v1/users/{id}/enable": {
      "post": {
        "tags": ["users"], "summary": "Enable a user (returns to active/waiting)", "security": [{ "BearerAuth": ["users.update"] }],
        "parameters": [{ "$ref": "#/components/parameters/UserID" }],
        "responses": { "200": { "description": "Enabled", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } } }
      }
    },
...
    "/api/v1/devices/{id}/enable": {
      "post": { "tags": ["devices"], "summary": "Enable a device", "security": [{ "BearerAuth": ["devices.write"] }], "parameters": [{ "$ref": "#/components/parameters/DeviceID" }], "responses": { "200": { "description": "Enabled" } } }
    },
    "/api/v1/devices/{id}/disable": {
      "post": { "tags": ["devices"], "summary": "Disable a device", "security": [{ "BearerAuth": ["devices.write"] }], "parameters": [{ "$ref": "#/components/parameters/DeviceID" }], "responses": { "200": { "description": "Disabled" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "description": "API token: `Authorization: Bearer wg_…` with scopes; separate from admin sessions." }
    },
    "parameters": {
...
      "UserID": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "Unauthorized": { "description": "Missing or invalid token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "Token lacks the required scope", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Conflict": { "description": "Conflict (exists, limit reached, pool exhausted, in use)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Invalid": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
...
      "UserStatus": { "type": "string", "enum": ["active", "disabled", "suspended", "expired", "traffic_exceeded", "waiting_first_connection"] },
...
      "UserPage": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
          "next_cursor": { "type": "string", "description": "Opaque; empty = last page" }
        }
      },
...
    }
  },
...
}