# internal/web/web.go
// Package web is the admin panel: server-rendered html/template pages with
// htmx partial swaps, session cookie auth, per-request CSRF tokens, and a
// strict content security policy. Handlers call domain services directly —
// the REST API is never in the request path (docs/architecture/
// project-structure.md: api/web → services).
//
// Security posture (docs/operations/security.md): session tokens are hashed
// at rest by internal/auth; the CSRF token is HMAC-derived from the session
// token so it needs no storage; scripts/styles are same-origin only (no
// inline script/style survive CSP); secrets never reach templates.
package web

import (
	"log/slog"
	"net/http"
	"strings"
	"sync"
	"time"

	"github.com/Sir-Adnan/wg-guard/internal/accounting"
	"github.com/Sir-Adnan/wg-guard/internal/admin"
	"github.com/Sir-Adnan/wg-guard/internal/audit"
	"github.com/Sir-Adnan/wg-guard/internal/auth"
	"github.com/Sir-Adnan/wg-guard/internal/backup"
	"github.com/Sir-Adnan/wg-guard/internal/cleanup"
	"github.com/Sir-Adnan/wg-guard/internal/clientconf"
	"github.com/Sir-Adnan/wg-guard/internal/config"
	"github.com/Sir-Adnan/wg-guard/internal/database"
	"github.com/Sir-Adnan/wg-guard/internal/device"
	"github.com/Sir-Adnan/wg-guard/internal/iface"
	"github.com/Sir-Adnan/wg-guard/internal/integration"
	"github.com/Sir-Adnan/wg-guard/internal/plan"
	"github.com/Sir-Adnan/wg-guard/internal/reseller"
	"github.com/Sir-Adnan/wg-guard/internal/secrets"
	"github.com/Sir-Adnan/wg-guard/internal/settings"
	"github.com/Sir-Adnan/wg-guard/internal/subscription"
	"github.com/Sir-Adnan/wg-guard/internal/telemetry"
	"github.com/Sir-Adnan/wg-guard/internal/token"
	"github.com/Sir-Adnan/wg-guard/internal/updatequeue"
	"github.com/Sir-Adnan/wg-guard/internal/user"
	"github.com/Sir-Adnan/wg-guard/internal/webhook"
)

// Deps wires the services the panel renders. The same instances the REST
// API uses are passed in — one business layer, two surfaces.
type Deps struct {
	DB          *database.DB
	Sessions    *auth.SessionStore
	Admins      *admin.Service
	Resellers   *reseller.Service
	Settings    *settings.Registry
	Ring        *secrets.KeyRing
	Audit       *audit.Service
	Users       *user.Service
	Devices     *device.Service
	Plans       *plan.Service
	Integration *integration.Service
	Ifaces      *iface.Service
	// ProfileGenerator is the canonical server-side profile preview seam.
	// It defaults to Ifaces.GenerateProfile; tests may replace it to exercise
	// entropy failures without weakening the production generator.
	ProfileGenerator func(iface.ProfilePolicy) (iface.Obfuscation, error)
	Accounting       *accounting.Service
	Log              *slog.Logger

	// Reconciler runs the complete serialized network pass after structural
	// mutations (see api.Server).
	Reconciler accounting.Reconciler

	// ClientConf renders client configs + QR (shared with the REST API).
	ClientConf *clientconf.Renderer

	// Links serves the per-user subscription links (public /sub/ surface).
	Links *subscription.Service

	// Backup is the archive engine (panel + CLI only — ADR-0007). Wired
	// from serve; nil in tests that don't exercise the ops screens.
	Backup *backup.Service
	// UpdateQueue is the data-volume side of the fixed host-owned lifecycle
	// broker. UpdateCatalog supplies published stable versions on demand.
	UpdateQueue   *updatequeue.Queue
	UpdateCatalog ReleaseCatalog

	// Tokens and Webhooks are the same instances the REST API uses — one
	// business layer, two surfaces.
	Tokens   *token.Service
	Webhooks *webhook.Service

	// Telemetry is the scheduler-owned immutable live history. Dashboard
	// requests never invoke its source.
	Telemetry *telemetry.Sampler

	Version      string
	TLSMode      config.TLSMode
	NodeID       string
	ToolsVersion string
}

// Server is the admin panel.
type Server struct {
	Deps
	assets          assetSet
	pages           map[string]*pageTemplate
	visualPresets   visualPresetCatalog
	visualPresetCSS []byte
	loginRL         *ipLimiter
	subRL           *ipLimiter // public /sub/ surface: request-rate window per IP
	updateCache     releaseCache
	cleanup         *cleanup.Service
	cleanupMu       sync.Mutex
}

// New builds the panel: parse templates once, hash assets once.
func New(d Deps) (*Server, error) {
	s := &Server{
		Deps:    d,
		loginRL: newIPLimiter(),
		subRL:   newRateLimiter(time.Minute, 60),
	}
	if s.ClientConf == nil {
		s.ClientConf = &clientconf.Renderer{
			Devices: d.Devices, Ifaces: d.Ifaces, Settings: d.Settings,
		}
	}
	if s.Resellers == nil && s.DB != nil {
		s.Resellers = reseller.NewService(s.DB)
	}
	if s.Integration == nil && s.DB != nil && s.Plans != nil {
		s.Integration = &integration.Service{DB: s.DB, Users: s.Users, Plans: s.Plans}
	}
	if s.ProfileGenerator == nil && d.Ifaces != nil {
		s.ProfileGenerator = d.Ifaces.GenerateProfile
	}
	if d.DB != nil {
		s.cleanup = cleanup.New(d.DB, d.Users, d.Ring)
	}
	var err error
	s.visualPresets, s.visualPresetCSS, err = loadVisualPresets()
	if err != nil {
		return nil, err
	}
	if err := s.initAssets(); err != nil {
		return nil, err
	}
	if err := s.initTemplates(); err != nil {
		return nil, err
	}
	return s, nil
}

// Handler returns the panel mux. It is mounted at "/" by serve; the REST
// API, health endpoints and /metrics keep their own root-level patterns.
func (s *Server) Handler() http.Handler {
	mux := http.NewServeMux()
	mux.HandleFunc("GET /", s.handleNotFound)
	mux.HandleFunc("GET /assets/", s.handleAssets)
	mux.HandleFunc("GET /cleanup", s.requirePermission(auth.ScopeCleanupManage, s.handleCleanupPage))
	mux.HandleFunc("POST /cleanup/preview", s.requirePermission(auth.ScopeCleanupManage, s.handleCleanupPreview))
	mux.HandleFunc("POST /cleanup/execute", s.requirePermission(auth.ScopeCleanupManage, s.handleCleanupExecute))
	mux.HandleFunc("POST /cleanup/optimize", s.requirePermission(auth.ScopeCleanupManage, s.handleCleanupOptimize))

	// --- auth (public) ---
	mux.HandleFunc("GET /login", s.handleLoginPage)
	mux.HandleFunc("POST /login", s.handleLoginSubmit)
	mux.HandleFunc("GET /onboarding", s.handleOnboardingPage)
	mux.HandleFunc("POST /onboarding", s.handleOnboardingSubmit)

	// --- preferences (session) ---
	mux.HandleFunc("POST /prefs/locale", s.requireSignedIn(s.handleLocaleSet))
	mux.HandleFunc("GET /appearance", s.requireSignedIn(s.handleAppearancePage))
	mux.HandleFunc("POST /appearance/me", s.requireSignedIn(s.handleAppearanceMe))
	mux.HandleFunc("POST /appearance/me/reset", s.requireSignedIn(s.handleAppearanceMeReset))
	mux.HandleFunc("POST /appearance/default", s.requireSignedIn(s.handleAppearanceDefault))
	mux.HandleFunc("POST /appearance/default/reset", s.requireSignedIn(s.handleAppearanceDefaultReset))
	mux.HandleFunc("POST /logout", s.requireSignedIn(s.handleLogout))

	// --- public subscription pages (token-gated, rate-limited) ---
	mux.HandleFunc("GET /sub/{token}", s.handleSubPage)
	mux.HandleFunc("GET /sub/{token}/devices/{deviceID}/qr", s.handleSubDeviceQR)
	mux.HandleFunc("GET /sub/{token}/devices/{deviceID}/config", s.handleSubDeviceConfig)

	// --- app pages ---
	mux.HandleFunc("GET /reseller/users", s.requireReseller(auth.ScopeUsersRead, s.handleResellerUsers))
	mux.HandleFunc("GET /reseller/users/{id}", s.requireReseller(auth.ScopeUsersRead, s.handleResellerUser))
	mux.HandleFunc("POST /reseller/users/{id}/next-plan", s.requireReseller(auth.ScopeNextPlansWrite, s.handleResellerNextPlanQueue))
	mux.HandleFunc("POST /reseller/users/{id}/next-plan/cancel", s.requireReseller(auth.ScopeNextPlansWrite, s.handleResellerNextPlanCancel))
	mux.HandleFunc("POST /reseller/users/{id}/traffic/reset", s.requireReseller(auth.ScopeTrafficUpdate, s.handleResellerTrafficReset))
	mux.HandleFunc("GET /reseller/devices/{id}/config", s.requireReseller(auth.ScopeConfigsRead, s.handleResellerDeviceConfig))
	mux.HandleFunc("GET /reseller/devices/{id}/qr", s.requireReseller(auth.ScopeConfigsRead, s.handleResellerDeviceQR))
	mux.HandleFunc("GET /reseller/tokens", s.requireReseller(auth.ScopeAPITokensManage, s.handleResellerTokensPage))
	mux.HandleFunc("POST /reseller/tokens", s.requireReseller(auth.ScopeAPITokensManage, s.handleResellerTokenCreate))
	mux.HandleFunc("POST /reseller/tokens/{tokenID}/revoke", s.requireReseller(auth.ScopeAPITokensManage, s.handleResellerTokenRevoke))
	mux.HandleFunc("GET /reseller/webhooks", s.requireReseller(auth.ScopeWebhooksRead, s.handleWebhooksPage))
	mux.HandleFunc("POST /reseller/webhooks/create", s.requireReseller(auth.ScopeWebhooksWrite, s.handleWebhookCreate))
	mux.HandleFunc("GET /reseller/webhooks/{id}", s.requireReseller(auth.ScopeWebhooksRead, s.handleWebhookShow))
	mux.HandleFunc("POST /reseller/webhooks/{id}/update", s.requireReseller(auth.ScopeWebhooksWrite, s.handleWebhookUpdate))
	mux.HandleFunc("POST /reseller/webhooks/{id}/rotate", s.requireReseller(auth.ScopeWebhooksWrite, s.handleWebhookRotate))
	mux.HandleFunc("POST /reseller/webhooks/{id}/delete", s.requireReseller(auth.ScopeWebhooksWrite, s.handleWebhookDelete))
	mux.HandleFunc("POST /reseller/webhooks/{id}/redeliver", s.requireReseller(auth.ScopeWebhooksWrite, s.handleWebhookRedeliver))
	mux.HandleFunc("GET /{$}", s.requireAuth(s.handleDashboard))
	mux.HandleFunc("GET /dashboard", s.requireAuth(s.handleDashboard))
	mux.HandleFunc("GET /dashboard/live", s.requireAuth(s.handleDashboardLive))
	mux.HandleFunc("GET /dashboard/chart", s.requireAuth(s.handleDashboardChart))
	mux.HandleFunc("GET /updates", s.requireMaintenanceRead(s.handleUpdatesPage))
	mux.HandleFunc("GET /updates/status", s.requireMaintenanceRead(s.handleUpdateStatus))
	mux.HandleFunc("POST /updates/request", s.requirePermission(auth.ScopeUpdateManage, s.handleUpdateRequest))
	mux.HandleFunc("POST /updates/cancel", s.requirePermission(auth.ScopeUpdateManage, s.handleUpdateCancel))
	mux.HandleFunc("POST /updates/snooze", s.requireMaintenanceRead(s.handleUpdateSnooze))
	mux.HandleFunc("GET /updates/report/{id}", s.requireMaintenanceRead(s.handleUpdateReport))
	mux.HandleFunc("GET /updates/backup", s.requirePermission(auth.ScopeBackupManage, s.handleUpdateBackup))

	// --- users ---
	mux.HandleFunc("GET /users/bulk", s.requirePermission(auth.ScopeUsersBulk, s.handleUserBulkPage))
	mux.HandleFunc("GET /users", s.requirePermission(auth.ScopeUsersRead, s.handleUserList))
	mux.HandleFunc("GET /users/new", s.requirePermission(auth.ScopeUsersCreate, s.handleUserNew))
	mux.HandleFunc("POST /users", s.requirePermission(auth.ScopeUsersCreate, s.handleUserCreate))
	mux.HandleFunc("POST /users/bulk", s.requirePermission(auth.ScopeUsersBulk, s.handleBulkCreate))
	mux.HandleFunc("POST /users/bulk-action", s.requirePermission(auth.ScopeUsersBulk, s.handleBulkAction))
	mux.HandleFunc("GET /users/{id}", s.requirePermission(auth.ScopeUsersRead, s.handleUserDetail))
	mux.HandleFunc("GET /users/{id}/edit", s.requirePermission(auth.ScopeUsersUpdate, s.handleUserEditPage))
	mux.HandleFunc("POST /users/{id}/edit", s.requirePermission(auth.ScopeUsersUpdate, s.handleUserUpdate))
	mux.HandleFunc("POST /users/{id}/enable", s.requirePermission(auth.ScopeUsersUpdate, s.handleUserEnable))
	mux.HandleFunc("POST /users/{id}/disable", s.requirePermission(auth.ScopeUsersUpdate, s.handleUserDisable))
	mux.HandleFunc("POST /users/{id}/delete", s.requirePermission(auth.ScopeUsersDelete, s.handleUserDelete))
	mux.HandleFunc("POST /users/{id}/restore", s.requirePermission(auth.ScopeUsersUpdate, s.handleUserRestore))
	mux.HandleFunc("POST /users/{id}/renew", s.requirePermission(auth.ScopeUsersUpdate, s.handleUserRenew))
	mux.HandleFunc("POST /users/{id}/quota/add", s.requirePermission(auth.ScopeUsersUpdate, s.handleUserQuotaAdd))
	mux.HandleFunc("POST /users/{id}/traffic/add", s.requirePermission(auth.ScopeTrafficUpdate, s.handleUserTrafficAdd)) // legacy meter correction
	mux.HandleFunc("POST /users/{id}/traffic/reset", s.requirePermission(auth.ScopeTrafficUpdate, s.handleUserTrafficReset))
	mux.HandleFunc("POST /users/{id}/next-plan", s.requirePermission(auth.ScopeNextPlansWrite, s.handleUserNextPlanQueue))
	mux.HandleFunc("POST /users/{id}/next-plan/cancel", s.requirePermission(auth.ScopeNextPlansWrite, s.handleUserNextPlanCancel))
	mux.HandleFunc("GET /users/{id}/configs.zip", s.requirePermission(auth.ScopeConfigsRead, s.handleUserConfigsArchive))
	mux.HandleFunc("POST /users/{id}/sub/create", s.requirePermission(auth.ScopeUsersUpdate, s.handleSubCreate))
	mux.HandleFunc("POST /users/{id}/sub/regenerate", s.requirePermission(auth.ScopeUsersUpdate, s.handleSubRegenerate))
	mux.HandleFunc("POST /users/{id}/sub/revoke", s.requirePermission(auth.ScopeUsersUpdate, s.handleSubRevoke))
	mux.HandleFunc("POST /users/{id}/sub/restore", s.requirePermission(auth.ScopeUsersUpdate, s.handleSubRestore))
	mux.HandleFunc("POST /users/{id}/devices", s.requirePermission(auth.ScopeDevicesWrite, s.handleDeviceCreate))

	// --- devices ---
	mux.HandleFunc("POST /devices/{id}/enable", s.requirePermission(auth.ScopeDevicesWrite, s.handleDeviceEnable))
	mux.HandleFunc("POST /devices/{id}/disable", s.requirePermission(auth.ScopeDevicesWrite, s.handleDeviceDisable))
	mux.HandleFunc("POST /devices/{id}/regenerate", s.requirePermission(auth.ScopeDevicesWrite, s.handleDeviceRegenerate))
	mux.HandleFunc("POST /devices/{id}/delete", s.requirePermission(auth.ScopeDevicesWrite, s.handleDeviceDelete))
	mux.HandleFunc("GET /devices/{id}/config", s.requirePermission(auth.ScopeConfigsRead, s.handleDeviceConfig))
	mux.HandleFunc("GET /devices/{id}/qr", s.requirePermission(auth.ScopeConfigsRead, s.handleDeviceQR))

	// --- plans ---
	mux.HandleFunc("GET /templates", s.requirePermission(auth.ScopeTemplatesRead, s.handlePlanList))
	mux.HandleFunc("GET /templates/new", s.requirePermission(auth.ScopeTemplatesWrite, s.handlePlanNew))
	mux.HandleFunc("POST /templates", s.requirePermission(auth.ScopeTemplatesWrite, s.handlePlanCreate))
	mux.HandleFunc("GET /templates/{id}/edit", s.requirePermission(auth.ScopeTemplatesWrite, s.handlePlanEditPage))
	mux.HandleFunc("POST /templates/{id}/edit", s.requirePermission(auth.ScopeTemplatesWrite, s.handlePlanUpdate))
	mux.HandleFunc("POST /templates/{id}/enable", s.requirePermission(auth.ScopeTemplatesWrite, s.handlePlanEnable))
	mux.HandleFunc("POST /templates/{id}/disable", s.requirePermission(auth.ScopeTemplatesWrite, s.handlePlanDisable))
	mux.HandleFunc("POST /templates/{id}/delete", s.requirePermission(auth.ScopeTemplatesWrite, s.handlePlanDelete))

	// --- interfaces ---
	mux.HandleFunc("GET /interfaces", s.requirePermission(auth.ScopeIfaceRead, s.handleIfaceList))
	mux.HandleFunc("GET /interfaces/new", s.requirePermission(auth.ScopeIfaceWrite, s.handleIfaceNew))
	mux.HandleFunc("POST /interfaces", s.requirePermission(auth.ScopeIfaceWrite, s.handleIfaceCreate))
	mux.HandleFunc("POST /interfaces/profile-preview", s.requirePermission(auth.ScopeIfaceWrite, s.handleProfilePreview))
	mux.HandleFunc("GET /interfaces/{id}/edit", s.requirePermission(auth.ScopeIfaceWrite, s.handleIfaceEditPage))
	mux.HandleFunc("POST /interfaces/{id}/edit", s.requirePermission(auth.ScopeIfaceWrite, s.handleIfaceUpdate))
	mux.HandleFunc("POST /interfaces/{id}/enable", s.requirePermission(auth.ScopeIfaceWrite, s.handleIfaceEnable))
	mux.HandleFunc("POST /interfaces/{id}/disable", s.requirePermission(auth.ScopeIfaceWrite, s.handleIfaceDisable))
	mux.HandleFunc("POST /interfaces/{id}/delete", s.requirePermission(auth.ScopeIfaceWrite, s.handleIfaceDelete))

	// --- settings (node.settings: the registry is operator territory) ---
	mux.HandleFunc("GET /settings", s.requirePermission(auth.ScopeNodeSettings, s.handleSettingsPage))
	mux.HandleFunc("POST /settings", s.requirePermission(auth.ScopeNodeSettings, s.handleSettingsSave))

	// --- backups (backup.manage; ADR-0007: panel/CLI only) ---
	mux.HandleFunc("GET /backups", s.requirePermission(auth.ScopeBackupManage, s.handleBackupsPage))
	mux.HandleFunc("POST /backups/create", s.requirePermission(auth.ScopeBackupManage, s.handleBackupCreate))
	mux.HandleFunc("POST /backups/import", s.requirePermission(auth.ScopeBackupManage, s.handleBackupImport))
	mux.HandleFunc("POST /backups/delete", s.requirePermission(auth.ScopeBackupManage, s.handleBackupDelete))
	mux.HandleFunc("GET /backups/{name}/download", s.requirePermission(auth.ScopeBackupManage, s.handleBackupDownload))
	mux.HandleFunc("POST /backups/restore", s.requirePermission(auth.ScopeBackupManage, s.handleBackupRestore))
	mux.HandleFunc("POST /backups/restore/confirm", s.requirePermission(auth.ScopeBackupManage, s.handleBackupRestoreConfirm))
	mux.HandleFunc("POST /backups/restore/cancel", s.requirePermission(auth.ScopeBackupManage, s.handleBackupRestoreCancel))
	mux.HandleFunc("POST /backups/schedules", s.requirePermission(auth.ScopeBackupManage, s.handleScheduleCreate))
	mux.HandleFunc("POST /backups/schedules/{id}/update", s.requirePermission(auth.ScopeBackupManage, s.handleScheduleUpdate))
	mux.HandleFunc("POST /backups/schedules/{id}/delete", s.requirePermission(auth.ScopeBackupManage, s.handleScheduleDelete))
	mux.HandleFunc("POST /backups/schedules/{id}/toggle", s.requirePermission(auth.ScopeBackupManage, s.handleScheduleToggle))
	mux.HandleFunc("POST /backups/telegram-test", s.requirePermission(auth.ScopeBackupManage, s.handleTelegramTest))

	// --- administrators (admins.manage) ---
	mux.HandleFunc("GET /resellers", s.requireOwner(s.handleResellersPage))
	mux.HandleFunc("POST /resellers", s.requireOwner(s.handleResellerCreate))
	mux.HandleFunc("POST /resellers/{id}/permissions", s.requireOwner(s.handleResellerPermissions))
	mux.HandleFunc("POST /resellers/{id}/templates", s.requireOwner(s.handleResellerTemplates))
	mux.HandleFunc("POST /resellers/{id}/enable", s.requireOwner(s.handleResellerEnable))
	mux.HandleFunc("POST /resellers/{id}/admins", s.requireOwner(s.handleResellerAdminCreate))
	mux.HandleFunc("GET /resellers/{id}/tokens", s.requireOwner(s.handleResellerTokensPage))
	mux.HandleFunc("POST /resellers/{id}/tokens", s.requireOwner(s.handleResellerTokenCreate))
	mux.HandleFunc("POST /resellers/{id}/tokens/{tokenID}/revoke", s.requireOwner(s.handleResellerTokenRevoke))
	mux.HandleFunc("GET /admins", s.requirePermission(auth.ScopeAdminsManage, s.handleAdminsPage))
	mux.HandleFunc("POST /admins/create", s.requirePermission(auth.ScopeAdminsManage, s.handleAdminCreate))
	mux.HandleFunc("POST /admins/{id}/password", s.requirePermission(auth.ScopeAdminsManage, s.handleAdminPassword))
	mux.HandleFunc("POST /admins/{id}/permissions", s.requirePermission(auth.ScopeAdminsManage, s.handleAdminPermissions))
	mux.HandleFunc("POST /admins/{id}/enable", s.requirePermission(auth.ScopeAdminsManage, s.handleAdminEnable))
	mux.HandleFunc("POST /admins/{id}/delete", s.requirePermission(auth.ScopeAdminsManage, s.handleAdminDelete))

	// --- API tokens (api_tokens.manage) ---
	mux.HandleFunc("GET /tokens", s.requirePermission(auth.ScopeAPITokensManage, s.handleTokensPage))
	mux.HandleFunc("POST /tokens/create", s.requirePermission(auth.ScopeAPITokensManage, s.handleTokenCreate))
	mux.HandleFunc("POST /tokens/{id}/revoke", s.requirePermission(auth.ScopeAPITokensManage, s.handleTokenRevoke))

	// --- webhooks (webhooks.read / webhooks.write) ---
	mux.HandleFunc("GET /webhooks", s.requirePermission(auth.ScopeWebhooksRead, s.handleWebhooksPage))
	mux.HandleFunc("POST /webhooks/create", s.requirePermission(auth.ScopeWebhooksWrite, s.handleWebhookCreate))
	mux.HandleFunc("GET /webhooks/{id}", s.requirePermission(auth.ScopeWebhooksRead, s.handleWebhookShow))
	mux.HandleFunc("POST /webhooks/{id}/update", s.requirePermission(auth.ScopeWebhooksWrite, s.handleWebhookUpdate))
	mux.HandleFunc("POST /webhooks/{id}/rotate", s.requirePermission(auth.ScopeWebhooksWrite, s.handleWebhookRotate))
	mux.HandleFunc("POST /webhooks/{id}/delete", s.requirePermission(auth.ScopeWebhooksWrite, s.handleWebhookDelete))
	mux.HandleFunc("POST /webhooks/{id}/redeliver", s.requirePermission(auth.ScopeWebhooksWrite, s.handleWebhookRedeliver))

	// --- audit (audit.view) ---
	mux.HandleFunc("GET /audit", s.requirePermission(auth.ScopeAuditView, s.handleAuditPage))

	h := http.Handler(mux)
	h = s.requireCSRF(h)
	h = s.sessionMiddleware(h)
	h = securityHeaders(h)
	h = bodyCap(h)
	return h
}

// The shared error surface is used for unmatched browser pages. Binary downloads,
// API envelopes and method-specific responses keep their existing contracts.
func (s *Server) handleNotFound(w http.ResponseWriter, r *http.Request) {
	if r.URL.Path == "/sub" || strings.HasPrefix(r.URL.Path, "/sub/") {
		s.surfaceError(w, r, http.StatusNotFound, "common.error_not_found", "sub")
		return
	}
	s.surfaceError(w, r, http.StatusNotFound, "common.error_not_found", "")
}